+++ programs/nft-voter/src/state/collection_config.rs
@@ -15,7 +15,9 @@ pub struct CollectionConfig {
 
     /// Governance power weight of the collection
     /// Each NFT in the collection has governance power = 1 * weight
-    pub weight: u16,
+    /// Note: The weight is scaled accordingly to the governing_token_mint decimals
+    /// Ex: if the the mint has 2 decimal places then weight of 1 should be stored as 100
+    pub weight: u64,
 
     /// Reserved for future upgrades
     pub reserved: [u8; 8],
@@ -23,8 +25,8 @@ pub struct CollectionConfig {
 
 impl CollectionConfig {
     pub fn get_max_weight(&self) -> u64 {
-        self.size
-            .checked_mul(self.weight as u32)
+        (self.size as u64)
+            .checked_mul(self.weight)
             .unwrap()
             .try_into()
             .unwrap()
